Removed unreferenced files
